# Performance improvements for UDDSketch