--- a/crates/udd-sketch/src/lib.rs
+++ b/crates/udd-sketch/src/lib.rs
@@ -81,14 +81,14 @@ impl SketchHashKey {
 }
 
 // Entries in the SketchHashMap contain a count and the next valid index of the map.
-#[derive(Serialize, Deserialize, Debug, Clone, PartialEq)]
+#[derive(Debug, Clone, PartialEq)]
 struct SketchHashEntry {
     count: u64,
     next: SketchHashKey,
 }
 
 // SketchHashMap is a special hash map of SketchHashKey->count that also keeps the equivalent of a linked list of the entries by increasing key value.
-#[derive(Serialize, Deserialize, Debug, Clone, PartialEq)]
+#[derive(Debug, Clone, PartialEq)]
 struct SketchHashMap {
     map: HashMap<SketchHashKey, SketchHashEntry>,
     head: SketchHashKey,
@@ -255,7 +255,7 @@ impl SketchHashMap {
     }
 }
 
-#[derive(Serialize, Deserialize, Clone, Debug, PartialEq)]
+#[derive(Clone, Debug, PartialEq)]
 pub struct UDDSketch {
     buckets: SketchHashMap,
     alpha: f64,