@feikesteenbergen 2025-04-10
There's probaby similar improvements in other parts of the code possible. However, I'd like to merge those kinds of fixes one by one myself.